Refactor registration and activation with COM
Refactor registration and activation with COM and provide one method for
setting the activation callback instead of handling each toast
separately. To distinguish between toasts the user can add unique
arguments for each toast but support for that isn't implemented yet.

With the refactoring comes also the ability to set the icon path and
background color in the notification itself and the windows settings
page for the associated app so methods for that were added.

One more benefit from this refactoring is the ability for the app to be
activated even if it's closed. The app will be launched with a
"-ToastActivated" argument which the app can use for quickly setting the
activation callback and then initialize the library.